Overhaul bash handling:
It took me a while to get to the root of this one: for some reason I
can't figure out, bash sometimes calls DEBUG with 2>/dev/null.

Because our hooks want to reliably show the user stderr, we want to skip
running the hooks when stderr is discarded in this fashion.

The simplest way to do this that I could think of is to check the device
number of /dev/fd/2, and if it's 3, which is /dev/null's device number
(on macOS, at least), skip this invocation.

I also cleaned up and better-documented the insanity we were already
doing to support bash and set -x