Unify bash and zsh a bit more:
Now that we've whittled away more of the cases where DEBUG is run, it
actually looks a lot like zsh's `preexec_functions`, so let's just
masquerade as that.

Drop the zsh `chpwd_functions` registration. This is completely
redundant with the hook firing for both `preexec` and `precmd`.

Also drop support for fish. The README documents why.