Refactor testing
The previous method of interacting with the shell didn't synchronize
sending commands with waiting for the prompt, which could result in
strange output where the next command would show up as part of the
output of the previous command, making the test flaky. This flakiness is
especially visible as the number of commands sent goes up.